🐛 Improve some stuff

+++ app/Http/Controllers/ClassroomUserController.php
@@ -2,8 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\ClassroomUser;
 use Illuminate\Http\Request;
+use App\Models\ClassroomUser;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class ClassroomUserController extends Controller
 {
@@ -40,7 +41,7 @@ class ClassroomUserController extends Controller
      */
     public function show(string $user_id, string $classroom_id)
     {
-        return ClassroomUser::where('user_id', '=', $user_id)->where('classroom_id', '=', $classroom_id)->get();
+        return ClassroomUser::where('user_id', '=', $user_id)->where('classroom_id', '=', $classroom_id)->first();
     }
 
     /**
@@ -66,12 +67,20 @@ class ClassroomUserController extends Controller
      */
     public function destroy(string $user_id, string $classroom_id)
     {
-        if (!$classroomUser = ClassroomUser::where('user_id', '=', $user_id)) {
-            return response()->json(['message' => 'Subscription not found'], 404);
-        }
-        $id = $classroomUser->first()->id;
-        if (ClassroomUser::destroy($id)) {
-            return response()->json(['message' => 'subscription deleted'], 200);
+        $user = JWTAuth::parseToken()->authenticate();
+        $classroomUserAdmin = ClassroomUser::where('user_id', '=', $user->id)->where('classroom_id', '=', $classroom_id)->first();
+        if ($classroomUserAdmin->role === 'admin' || $classroomUserAdmin->id == $user_id) {
+            if (!$classroomUser = ClassroomUser::where('user_id', '=', $user_id)) {
+                return response()->json(['message' => 'Subscription not found'], 404);
+            }
+            $id = $classroomUser->first()->id;
+            if (ClassroomUser::destroy($id)) {
+                return response()->json(['message' => 'subscription deleted'], 200);
+            } else {
+                return response()->json(['message' => 'An error occoured during operation'], 400);
+            }
+        } else {
+            return response()->json(['message' => 'Not authorized'], 403);
         }
     }
 }
